// packages/vvs-frontend/src/config/constants/types.ts
export type PageMeta = {
  title?: string
  description?: string
  image?: string
  canonical?: string
  ogImageUrl?: string
  siteName?: string
}